+++ packages/synapse/src/react/hooks/index.ts
@@ -1,3 +1,5 @@
+export * from './useApiMutation'
+export * from './useApiQuery'
 export * from './useCreateStorage'
 export * from './useObservable'
 export * from './useSelector'